ADD the option to manage profiles from the profiles button
